refactor: 関数を他のファイルに切り出したりフォルダ名を変更
