+++ src/features/stayLogGraph/components/SimulataneousStay.tsx
